Change: move the interfactes introduced in the Account Adder controller to the account and keystore interfaces
