+++ src/controllers/main/main.test.ts
@@ -1,4 +1,3 @@
-import { SelectedAccount } from 'controllers/accountAdder/accountAdder'
 import { ethers } from 'ethers'
 import fetch from 'node-fetch'
 
@@ -7,6 +6,7 @@ import { describe, expect, test } from '@jest/globals'
 import { produceMemoryStore } from '../../../test/helpers'
 import { AMBIRE_ACCOUNT_FACTORY } from '../../consts/deploy'
 import { BIP44_STANDARD_DERIVATION_TEMPLATE } from '../../consts/derivation'
+import { SelectedAccount } from '../../interfaces/account'
 import { UserRequest } from '../../interfaces/userRequest'
 import { KeyIterator } from '../../libs/keyIterator/keyIterator'
 import { KeystoreSigner } from '../../libs/keystoreSigner/keystoreSigner'
