+++ src/interfaces/account.ts
@@ -1,3 +1,5 @@
+import { NetworkDescriptor } from './networkDescriptor'
+
 export type AccountId = string
 
 export interface Account {
@@ -39,4 +41,57 @@ export type AccountStates = {
   [accountId: string]: {
     [networkId: string]: AccountOnchainState
   }
+}
+
+type AccountDerivationMeta = {
+  slot: number // the iteration on which the account is derived, starting from 1
+  index: number // the derivation index of the <account> in the slot, starting from 0
+  isLinked: boolean // linked accounts are also smart accounts, so use a flag to differentiate
+}
+
+export type AccountWithNetworkMeta = Account & { usedOnNetworks: NetworkDescriptor[] }
+
+/**
+ * The account that is derived programmatically and internally by Ambire.
+ * Could be either a basic (EOA) account, a derived with custom derivation
+ * basic (EOA) account (used for smart account key only) or a smart account.
+ */
+export type DerivedAccount = AccountDerivationMeta & { account: AccountWithNetworkMeta }
+// Sub-type, used during intermediate step during the deriving accounts process
+export type DerivedAccountWithoutNetworkMeta = Omit<DerivedAccount, 'account'> & {
+  account: Account
+}
+
+/**
+ * Enum for tracking the import status of an account during the import process.
+ */
+export enum ImportStatus {
+  NotImported = 'not-imported',
+  ImportedWithoutKey = 'imported-without-key', // as a view only account
+  ImportedWithSomeOfTheKeys = 'imported-with-some-of-the-keys', // imported with
+  // some of the keys (having the same key type), but not all found on the current page
+  ImportedWithTheSameKeys = 'imported-with-the-same-keys', // imported with all
+  // keys (having the same key type) found on the current page
+  ImportedWithDifferentKeys = 'imported-with-different-keys' // different key
+  // meaning that could be a key with the same address but different type,
+  // or a key with different address altogether.
+}
+
+/**
+ * All the accounts that should be visible on the current page - the Basic
+ * Accounts, Smart Accounts and the linked accounts. Excludes the derived
+ * EOA (basic) accounts used for smart account keys only.
+ */
+export type AccountOnPage = DerivedAccount & { importStatus: ImportStatus }
+
+/**
+ * The account that the user has actively chosen (selected) via the app UI.
+ * It's always one of the visible accounts returned by the accountsOnPage().
+ * Could be either a basic (EOA) account, a smart account or a linked account.
+ */
+export type SelectedAccount = {
+  account: Account
+  isLinked: AccountDerivationMeta['isLinked']
+  importStatus: ImportStatus
+  accountKeys: (Omit<AccountDerivationMeta, 'isLinked'> & { addr: Account['addr'] })[]
 }
